formatting json response for api

+++ src/main/java/com/example/assignment_3/controller/ProductController.java
@@ -5,6 +5,8 @@ import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -28,24 +30,34 @@ public class ProductController {
 
   // get product api
   @GetMapping("")
-  public List<Product> getProducts() {
+  public ResponseEntity<List<Product>> getProducts() {
     List<Product> products = productRepository.findAll();
-    logger.info("get products success");
 
-    return products;
+    logger.info("Get Products Success");
+    return new ResponseEntity<List<Product>>(products, HttpStatus.OK);
   }
 
   // create product api
   @PostMapping("")
-  public Product createProduct(@RequestBody Product data) {
-    logger.info("save product success");
-    return productRepository.save(data);
+  public ResponseEntity<Product> createProduct(@RequestBody Product data) {
+    Product product = productRepository.save(data);
+
+    logger.info("Create Product Success");
+    return new ResponseEntity<Product>(product, HttpStatus.CREATED);
   }
 
   // delete product api
   @DeleteMapping("{id}")
-  public void deleteProduct(@PathVariable Long id) {
+  public ResponseEntity<String> deleteProduct(@PathVariable Long id) {
+    // error handling for not found products
+    if (!productRepository.existsById(id)) {
+      logger.info("Product Doesn't Exist");
+      return new ResponseEntity<String>("Data Not Found", HttpStatus.NOT_FOUND);
+    }
+
     productRepository.deleteById(id);
-    logger.info("delete product success");
+
+    logger.info("Delete Product Success");
+    return new ResponseEntity<String>("Delete Success", HttpStatus.OK);
   }
 }
